refactor: move settings store file
