+++ src/apps/photo-companion/src/lib/layout/app/app.layout.svelte
@@ -5,7 +5,7 @@
 	import { version } from "@lib/constants";
 	import { iconGithub, iconQuestion, iconSettings } from "@lib/icons";
 	import { routeAbout, routeChangelog, routeSettings, urlGithub } from "@lib/routes";
-	import { settingsStore } from "@lib/settings-store";
+	import { settingsStore } from "@lib/stores";
 	import { Navigation } from "@lib/layout";
 	import { title } from "@lib/constants";
 	import styles from "./app.layout.module.css";
