+++ src/apps/photo-companion/src/lib/stores/index.ts
@@ -1,0 +1,1 @@
+export { settingsStore, SETTINGS_DEFAULT, type SettingsStore } from "./settings";
