+++ src/apps/photo-companion/src/lib/stores/settings/index.ts
@@ -1,4 +1,4 @@
-import { persistable } from "./helpers/persistable";
+import { persistable } from "../../helpers/persistable";
 import type {
 	BridgeEventName,
 	LightsCity,
