+++ src/apps/photo-companion/src/pages/settings/settings.const.ts
@@ -3,7 +3,7 @@ import { lightsCityList } from "@lib/constants";
 import { SUPPORTED_BRIDGES_NAME_SET, isNavigationTime } from "bridge-schedule";
 import { dict } from "@lib/dict";
 import type { InputSelectOption, Option } from "@lib/components";
-import type { SettingsStore } from "@lib/settings-store";
+import type { SettingsStore } from "@lib/stores/settings";
 import { isNullable } from "@shared/utils";
 
 export const BRIDGES_SPB_OPTIONS: Option[] = [
