+++ src/apps/photo-companion/src/services/lights.ts
@@ -1,5 +1,4 @@
-import { SETTINGS_DEFAULT } from "@lib/settings-store";
-import { settingsStore } from "@lib/settings-store";
+import { settingsStore, SETTINGS_DEFAULT } from "@lib/stores";
 import { initLightsProvider } from "lights-schedule";
 
 let city = SETTINGS_DEFAULT.lights_city;
